applications: nrf5340_audio: power: make some APIs internal
power_module_data_get, power_module_measurement_start and
power_module_measurement_stop are only used within the module, so
there's no need to expose them publicly.
