+++ applications/nrf5340_audio/src/modules/power_module.h
@@ -7,58 +7,6 @@
 #ifndef _POWER_MODULE_H_
 #define _POWER_MODULE_H_
 
-#include <stdint.h>
-
-typedef void (*nrf_power_module_handler_t)(uint8_t rail);
-
-/*
- * Different rails to be measured by the INA231
- * VBAT		= Battery
- * VDD1_CODEC	= 1.2V rail for CS47L63
- * VDD2_CODEC	= 1.8V rail for CS47L63
- * VDD2_NRF	= NRF5340
- */
-enum ina_name {
-	VBAT,
-	VDD1_CODEC,
-	VDD2_CODEC,
-	VDD2_NRF,
-};
-
-struct power_module_data {
-	float current;
-	float power;
-	float bus_voltage;
-	float shunt_voltage;
-};
-
-/**@brief   Read the latest measurements from a given INA231
- *
- * @param   name    Name of the INA231 to be read from
- * @param   data    Container for the read data
- */
-void power_module_data_get(enum ina_name name, struct power_module_data *data);
-
-/**@brief   Write configuration to INA231 to start measurements
- *
- * @param   name    Name of the INA231 to start
- *
- * @param   data_handler    Callback to use when data is ready, if set to NULL
- *			    the default callback for printing measurements will
- *			    be used
- *
- * @return  0 if successful
- */
-int power_module_measurement_start(enum ina_name name, nrf_power_module_handler_t data_handler);
-
-/**@brief   Stop continuous measurements from given INA231
- *
- * @param   name    Name of the INA231 to be stopped
- *
- * @return  0 if successful
- */
-int power_module_measurement_stop(enum ina_name name);
-
 /**@brief   Initialize power module
  *
  * @return  0 if successful
